renames package 'se.liu.ida.hefquin.engine.vocabulary' to 'se.liu.ida.hefquin.vocabulary'
