+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/vocabulary/ECVocab.java
@@ -1,4 +1,4 @@
-package se.liu.ida.hefquin.engine.vocabulary;
+package se.liu.ida.hefquin.vocabulary;
 
 /* CVS $Id: $ */
